% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{argH1}
\alias{argH1}
\title{Compute best path with one segment on the interval l:r}
\usage{
argH1(l, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Change point (\code{NULL}), state value and H1-value
}
\description{
Compute best path with one segment on the interval l:r
}
