% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{argH2}
\alias{argH2}
\title{Compute best path with two segments and jump at k1 on the interval l:r}
\usage{
argH2(l, k1, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{Change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Change point (\code{k1}), state values and H2-value
}
\description{
Compute best path with two segments and jump at k1 on the interval l:r
}
