% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{argHD_wrap}
\alias{argHD_wrap}
\title{Dispatch to argH1, argH2, or argH3 depending on the change-point
configuration}
\usage{
argHD_wrap(l, k1, k2, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{First candidate change point}

\item{k2}{Second candidate change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
List with k_star, i_star, h_star
}
\description{
Dispatch to argH1, argH2, or argH3 depending on the change-point
configuration
}
\keyword{internal}
