% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compare_norm}
\alias{compare_norm}
\title{Generate a sample and estimate paths using QATS, PMAP and Viterbi}
\usage{
compare_norm(n, m, Pi, pp, mu, sigma, d0, n_seeds, rotate, n_rep, n_sim)
}
\arguments{
\item{n}{Length of the sequence}

\item{m}{Cardinality of the state space}

\item{Pi}{Initial state distribution}

\item{pp}{Transition matrix}

\item{mu}{Means of the normal emission distributions}

\item{sigma}{Standard deviations of the normal emission distributions}

\item{d0}{Smallest search interval}

\item{n_seeds}{Number of seeds for the optimistic search}

\item{rotate}{Indicates whether or not the gain functions have to be rotated}

\item{n_rep}{Number of repetitions (for timing)}

\item{n_sim}{Number of simulations}
}
\value{
A list containing estimation times and errors for QATS, Viterbi,
and PMAP
}
\description{
Generate a sample and estimate paths using QATS, PMAP and Viterbi
}
