% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display.mat}
\alias{display.mat}
\title{Plot of the gain function for path with 3 segments}
\usage{
display.mat(
  l,
  r,
  x0,
  par,
  include.dim1 = FALSE,
  include.dim0 = FALSE,
  display = TRUE
)
}
\arguments{
\item{l}{Left-most index}

\item{r}{Right-most index}

\item{x0}{Previous state}

\item{par}{Model parameters}

\item{include.dim1}{If true, add the single change point score}

\item{include.dim0}{If true, add the no change point score}

\item{display}{Whether or not to display the map}
}
\value{
Plot
}
\description{
Plot of the gain function for path with 3 segments
}
