% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display.result}
\alias{display.result}
\title{Display final results}
\usage{
display.result(xx.0, xx.1 = NULL, xx.2 = NULL, par, yy = NULL)
}
\arguments{
\item{xx.0}{True hidden path}

\item{xx.1}{1st estimated path}

\item{xx.2}{2nd estimated path}

\item{par}{Model parameters}

\item{yy}{The observation sequence}
}
\value{
A ggplot object (also printed)
}
\description{
Display final results
}
