% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display.vec}
\alias{display.vec}
\title{Plot of the gain function for path with 2 segments}
\usage{
display.vec(l, r, x0, par, display = TRUE)
}
\arguments{
\item{l}{Left-most index}

\item{r}{Right-most index}

\item{x0}{Previous state}

\item{par}{Model parameters}

\item{display}{Whether or not to display the map}
}
\value{
Plot
}
\description{
Plot of the gain function for path with 2 segments
}
