% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{.validate_opts}
\alias{.validate_opts}
\title{Validate and fill in default QATS options}
\usage{
.validate_opts(opts, par)
}
\arguments{
\item{opts}{A list of options (partially or fully specified).}

\item{par}{The HMM parameter list (used only to check SS bounds).}
}
\value{
A fully validated and completed opts list.
}
\description{
Validate and fill in default QATS options
}
\keyword{internal}
