% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{G0}
\alias{G0}
\title{Log-probability of a hidden path}
\usage{
G0(xx, par)
}
\arguments{
\item{xx}{Vector of hidden states}

\item{par}{Model parameters}
}
\value{
Scalar log-probability
}
\description{
Log-probability of a hidden path
}
\keyword{internal}
