% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{G1}
\alias{G1}
\title{Score of a one-segment path on an interval}
\usage{
G1(l, r, x0, i1, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{i1}{State of the segment}

\item{par}{Model parameters}
}
\value{
Scalar score
}
\description{
Score of a one-segment path on an interval
}
\keyword{internal}
