% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{G2}
\alias{G2}
\title{Score of a two-segment path on an interval}
\usage{
G2(l, k1, r, x0, i1, i2, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{Change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{i1}{State of the first segment}

\item{i2}{State of the second segment}

\item{par}{Model parameters}
}
\value{
Scalar score
}
\description{
Score of a two-segment path on an interval
}
\keyword{internal}
