% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{G3}
\alias{G3}
\title{Score of a three-segment path on an interval}
\usage{
G3(l, k1, k2, r, x0, i1, i2, i3, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{First change point}

\item{k2}{Second change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{i1}{State of the first segment}

\item{i2}{State of the second segment}

\item{i3}{State of the third segment}

\item{par}{Model parameters}
}
\value{
Scalar score
}
\description{
Score of a three-segment path on an interval
}
\keyword{internal}
