% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{G_classifier.CPP}
\alias{G_classifier.CPP}
\title{G-classifier, C++ version}
\usage{
G_classifier.CPP(par, C1 = 0, C2 = 1, C3 = 0, C4 = 0)
}
\arguments{
\item{par}{Model parameters}

\item{C1}{Constant 1}

\item{C2}{Constant 2}

\item{C3}{Constant 3}

\item{C4}{Constant 4}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
G-classifier, C++ version
}
\details{
Special cases depending on the constants:
\itemize{
\item C1 = C3 = C4 = 0: Viterbi
\item C2 = C3 = C4 = 0: pointwise MAP
\item C1 = C2 = C3 = 0: maximum prior probability
\item C1 = C2 = C4 = 0: marginal prior mode
\item C1 = C3 = 0: generalized Viterbi
}
}
