% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_classifier_cpp}
\alias{G_classifier_cpp}
\title{G-classifier}
\usage{
G_classifier_cpp(C1, C2, C3, C4, n, m, Pi, logPi, pp, qq, f_mseq, g_mseq)
}
\arguments{
\item{C1}{Constant 1}

\item{C2}{Constant 2}

\item{C3}{Constant 3}

\item{C4}{Constant 4}

\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{Pi}{Initial distribution}

\item{logPi}{Initial log-distribution}

\item{pp}{Transition matrix}

\item{qq}{Log-transition matrix}

\item{f_mseq}{Densities of the emission distributions}

\item{g_mseq}{Log-densities of the emission distributions}
}
\value{
Estimated sequence
}
\description{
G-classifier
}
\keyword{internal}
