% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{H1}
\alias{H1}
\title{Best one-segment score on an interval, maximised over states}
\usage{
H1(l, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Scalar best score
}
\description{
Best one-segment score on an interval, maximised over states
}
\keyword{internal}
