% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{H1_dbl}
\alias{H1_dbl}
\title{Gain function for path with 1 segment}
\usage{
H1_dbl(D, s, x0, par)
}
\arguments{
\item{D}{Vector of indices}

\item{s}{Length of D}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Scalar representing the gain function on D
}
\description{
Gain function for path with 1 segment
}
\keyword{internal}
