% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{H2}
\alias{H2}
\title{Best two-segment score on an interval, maximised over states}
\usage{
H2(l, k1, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{Change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Scalar best score
}
\description{
Best two-segment score on an interval, maximised over states
}
\keyword{internal}
