% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{H2_vec}
\alias{H2_vec}
\title{Gain function for path with 2 segments}
\usage{
H2_vec(D, s, x0, par)
}
\arguments{
\item{D}{Vector of indices}

\item{s}{Length of D}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Vector representing the gain function on D
}
\description{
Gain function for path with 2 segments
}
\keyword{internal}
