% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob.R
\name{H3}
\alias{H3}
\title{Best three-segment score on an interval, maximised over states}
\usage{
H3(l, k1, k2, r, x0, par)
}
\arguments{
\item{l}{Left endpoint of the interval}

\item{k1}{First change point}

\item{k2}{Second change point}

\item{r}{Right endpoint of the interval}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Scalar best score
}
\description{
Best three-segment score on an interval, maximised over states
}
\keyword{internal}
