% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{H3_mat}
\alias{H3_mat}
\title{Gain function for path with 3 segments}
\usage{
H3_mat(D, s, x0, par)
}
\arguments{
\item{D}{Vector of indices}

\item{s}{Length of D}

\item{x0}{Previous state}

\item{par}{Model parameters}
}
\value{
Matrix representing the gain function on D
}
\description{
Gain function for path with 3 segments
}
\keyword{internal}
