% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{K_segmentation.CPP}
\alias{K_segmentation.CPP}
\title{K-segmentation, C++ version}
\usage{
K_segmentation.CPP(par, K_max = NULL)
}
\arguments{
\item{par}{Model parameters}

\item{K_max}{Maximum number of constant pieces}
}
\value{
Estimated sequences, their probabilities and estimation time
}
\description{
K-segmentation, C++ version
}
