% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{K_segmentation_cpp}
\alias{K_segmentation_cpp}
\title{K-segmentation}
\usage{
K_segmentation_cpp(K_max, n, m, logPi, qq, g_mseq)
}
\arguments{
\item{K_max}{Maximum number of constant pieces}

\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{logPi}{Initial log-distribution}

\item{qq}{Log-transition matrix}

\item{g_mseq}{Log-densities of the emission distributions}
}
\value{
Estimated sequences
}
\description{
K-segmentation
}
\keyword{internal}
