% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{lp_norm}
\alias{lp_norm}
\title{lp norm}
\usage{
lp_norm(xx.0, xx.1, p)
}
\arguments{
\item{xx.0}{Reference vector}

\item{xx.1}{Estimated vector}

\item{p}{Number indicating which l-norm should be computed}
}
\value{
lp-norm of xx.0 - xx.1
}
\description{
lp norm
}
