% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{OSH3}
\alias{OSH3}
\title{2-dimensional optimistic search}
\usage{
OSH3(l, r, x0, par, opts)
}
\arguments{
\item{l}{Left-most index}

\item{r}{Right-most index}

\item{x0}{Previous state}

\item{par}{Model parameters}

\item{opts}{A list of options for the estimation}
}
\value{
Triplet
}
\description{
2-dimensional optimistic search
}
\keyword{internal}
