% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{OSH3_k0}
\alias{OSH3_k0}
\title{2-dimensional optimistic search for H3 with a starting point}
\usage{
OSH3_k0(k0, l, r, x0, par, opts)
}
\description{
2-dimensional optimistic search for H3 with a starting point
}
\keyword{internal}
