% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PMAP_cpp}
\alias{PMAP_cpp}
\title{PMAP-classifier}
\usage{
PMAP_cpp(
  xx,
  alpha_hat,
  alpha_bar,
  beta_hat,
  beta_bar,
  cc_inv,
  n,
  m,
  Pi,
  pp,
  f_mseq
)
}
\arguments{
\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{Pi}{Initial distribution}

\item{pp}{Transition matrix}

\item{f_mseq}{Densities of the emission distributions}
}
\value{
Estimated sequence
}
\description{
PMAP-classifier
}
\keyword{internal}
