% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PMAP_timer_cpp}
\alias{PMAP_timer_cpp}
\title{PMAP-classifier, timing in C++}
\usage{
PMAP_timer_cpp(n_rep, n, m, Pi, pp, f_mseq)
}
\arguments{
\item{n_rep}{Number of repetitions (for timing)}

\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{Pi}{Initial distribution}

\item{pp}{Transition matrix}

\item{f_mseq}{Densities of the emission distributions}
}
\value{
Estimated sequence
}
\description{
PMAP-classifier, timing in C++
}
\keyword{internal}
