% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{QATS_timer_cpp}
\alias{QATS_timer_cpp}
\title{Quick Adaptive Ternary Segmentation, timing in C++}
\usage{
QATS_timer_cpp(d0, n_seeds, rotate, n_rep, n, m, logPi, qq, GG, SS, UU)
}
\arguments{
\item{d0}{Smallest search interval}

\item{n_seeds}{Number of seeds for the optimistic search}

\item{rotate}{Indicates whether or not the gain functions have to be rotated}

\item{n_rep}{Number of repetition (for timing)}

\item{n}{Length of the sequence}

\item{m}{Cardinality of the state space}

\item{logPi}{Log-initial distribution}

\item{qq}{Log-transition probabilities}

\item{GG}{Cumulative log-emission densities}

\item{SS}{Initial partition}

\item{UU}{Length of the initial partition}
}
\description{
Quick Adaptive Ternary Segmentation, timing in C++
}
\keyword{internal}
