% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sample.t.HMM}
\alias{sample.t.HMM}
\title{Generates an HMM process with student noise and precomputes parameters for
estimation}
\usage{
sample.t.HMM(n, m, Pi, pp, nu, mu, sigma)
}
\arguments{
\item{n}{Length of the sequence to generate}

\item{m}{Cardinality of the state space}

\item{Pi}{Initial state distribution}

\item{pp}{Transition matrix}

\item{nu}{Degrees of freedom}

\item{mu}{Mean vector}

\item{sigma}{Scale parameter}
}
\value{
The hidden sequence and the sequence of observations
}
\description{
Generates an HMM process with student noise and precomputes parameters for
estimation
}
\keyword{internal}
