% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{set.par}
\alias{set.par}
\title{Precomputes parameters for the estimation}
\usage{
set.par(
  yy,
  Pi,
  pp,
  emi.dist = "normal",
  emi.param = list(mu = 1:nrow(pp), sigma = rep(1, nrow(pp)))
)
}
\arguments{
\item{yy}{Sequence of observations}

\item{Pi}{Initial distribution}

\item{pp}{Transition matrix}

\item{emi.dist}{Emission distribution}

\item{emi.param}{Parameters for the emission distribution}
}
\value{
A list of computed parameters: log-probabilities and cumulative sums
}
\description{
Precomputes parameters for the estimation
}
