% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{SS.zz_xx}
\alias{SS.zz_xx}
\title{Transform a sequence into the partition it creates and its levels}
\usage{
SS.zz_xx(xx, n)
}
\arguments{
\item{xx}{Sequence of hidden states}

\item{n}{Length of the sequence}
}
\value{
Partition and corresponding levels
}
\description{
Transform a sequence into the partition it creates and its levels
}
\keyword{internal}
