% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{V_measure}
\alias{V_measure}
\title{V-measure}
\usage{
V_measure(xx.0, xx.1)
}
\arguments{
\item{xx.0}{Reference vector}

\item{xx.1}{Estimated vector}
}
\value{
V-measure of xx.0 and xx.1
}
\description{
V-measure
}
