% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Viterbi_cpp}
\alias{Viterbi_cpp}
\title{Viterbi decoder}
\usage{
Viterbi_cpp(xx, zeta, rho, n, m, logPi, qq, g_mseq)
}
\arguments{
\item{xx}{Pre-declared sequence to be updated}

\item{zeta}{Back-pointer matrix (m x n)}

\item{rho}{Score matrix (m x n)}

\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{logPi}{Initial log-distribution}

\item{qq}{Log-transition matrix}

\item{g_mseq}{Log-densities of the emission distributions}
}
\value{
Updates xx, zeta, and rho in place
}
\description{
Viterbi decoder
}
\keyword{internal}
