% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Viterbi_timer_cpp}
\alias{Viterbi_timer_cpp}
\title{Viterbi decoder, timing in C++}
\usage{
Viterbi_timer_cpp(n_rep, n, m, logPi, qq, g_mseq)
}
\arguments{
\item{n_rep}{Number of repetitions (for timing)}

\item{n}{Length of the observation sequence}

\item{m}{Cardinality of the state space}

\item{logPi}{Initial log-distribution}

\item{qq}{Log-transition matrix}

\item{g_mseq}{Log-densities of the emission distributions}
}
\value{
Estimated sequence
}
\description{
Viterbi decoder, timing in C++
}
\keyword{internal}
