% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{xx_SS.zz}
\alias{xx_SS.zz}
\title{Transform a partition and its levels to a sequence}
\usage{
xx_SS.zz(SS, zz)
}
\arguments{
\item{SS}{Partition in the form of a 2-by-U matrix}

\item{zz}{Vector with the UU different levels}
}
\value{
Corresponding vector of hidden states
}
\description{
Transform a partition and its levels to a sequence
}
\keyword{internal}
