% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display.0}
\alias{display.0}
\title{Displays the current fit along with the data}
\usage{
display.0(xx.0, xx.1 = NULL, xx.2 = NULL, par, SS = NULL, k_new = NULL)
}
\arguments{
\item{xx.0}{True hidden path}

\item{xx.1}{1st estimated path}

\item{xx.2}{2nd 2stimated path}

\item{par}{Model parameters}

\item{SS}{Partition}

\item{k_new}{Newly added change points}
}
\value{
Plot
}
\description{
Displays the current fit along with the data
}
\keyword{internal}
