% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display.1}
\alias{display.1}
\title{Displays the current fit and the gain functions, along with the data}
\usage{
display.1(
  xx.0,
  xx.1 = NULL,
  xx.2 = NULL,
  par,
  SS = NULL,
  k_new = NULL,
  CP = NULL,
  l = 1,
  res1 = NULL,
  t_star1 = NULL,
  k_star1 = NULL,
  res2 = NULL,
  t_star2 = NULL,
  k_star2 = NULL
)
}
\arguments{
\item{xx.0}{True hidden path}

\item{xx.1}{1st estimated path}

\item{xx.2}{2nd 2stimated path}

\item{par}{Model parameters}

\item{SS}{Partition}

\item{k_new}{Newly added change points}

\item{CP}{True change points}

\item{l}{Left-most index of the current window}

\item{res1}{Vector of 1-dimensional gain}

\item{t_star1}{True maximum}

\item{k_star1}{Maximum found}

\item{res2}{Matrix of 2-dimensional gain}

\item{t_star2}{True maxium}

\item{k_star2}{Maximum found}
}
\value{
Plot
}
\description{
Displays the current fit and the gain functions, along with the data
}
\keyword{internal}
