% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiors.R
\name{G_classifier.R}
\alias{G_classifier.R}
\title{G-classifier, R version}
\usage{
G_classifier.R(par, C1 = 0, C2 = 1, C3 = 0, C4 = 0)
}
\arguments{
\item{par}{Model parameters}

\item{C1}{Constant 1}

\item{C2}{Constant 2}

\item{C3}{Constant 3}

\item{C4}{Constant 4}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
G-classifier, R version
}
