% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{OS}
\alias{OS}
\title{Basic one-dimensional optimistic search}
\usage{
OS(L, R, d0, fun, argfun, M = -1)
}
\arguments{
\item{L}{Initial left index}

\item{R}{Initial right index}

\item{d0}{Smallest search interval}

\item{fun}{Function to optimize}

\item{argfun}{Function to optimize which also returns arguments}

\item{M}{Initial probe (optional)}
}
\value{
Triplet
}
\description{
Basic one-dimensional optimistic search
}
\keyword{internal}
