% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{OSH3_k0}
\alias{OSH3_k0}
\title{2-dimensional optimistic search, with starting point k0}
\usage{
OSH3_k0(k0, l, r, x0, par, opts)
}
\arguments{
\item{k0}{Starting point}

\item{l}{Left-most index}

\item{r}{Right-most index}

\item{x0}{Previous state}

\item{par}{Model parameters}

\item{opts}{A list of options for the estimation}
}
\value{
Triplet
}
\description{
2-dimensional optimistic search, with starting point k0
}
\keyword{internal}
