% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiors.R
\name{PMAP.CPP}
\alias{PMAP.CPP}
\title{PMAP decoder, C++ version}
\usage{
PMAP.CPP(par, n.rep = 1)
}
\arguments{
\item{par}{Model parameters}

\item{n.rep}{Number of repetition (for timing)}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
PMAP decoder, C++ version
}
