% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{QATS.CPP}
\alias{QATS.CPP}
\title{Quick Adaptive Ternary Segmentation, C++ version}
\usage{
QATS.CPP(
  par,
  opts = list(n.seeds = 3, d0 = 3, SS = NULL, rotate = FALSE, n.rep = 1)
)
}
\arguments{
\item{par}{Model parameters}

\item{opts}{A list of options for the estimation. Default is \code{opts =
list(n.seeds = 3, d0 = 3, rotate = FALSE, n.rep = 1)} where
\code{n.seeds} is the number of seeds for the optimistic search,
\code{d0} is the size of the smallest search interval,
\code{SS} is the initial partition,
\code{rotate} indicates whether or not the gain functions have to be rotated,
\code{n.rep} is the umber of repetitions (for timing).}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
Quick Adaptive Ternary Segmentation, C++ version
}
