% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{QATS.display}
\alias{QATS.display}
\title{Quick Adaptive Ternary Segmentation, with display}
\usage{
QATS.display(
  xx.0,
  xx.Vit,
  par,
  opts = list(n.seeds = 3, d0 = 3, SS = NULL, rotate = FALSE)
)
}
\arguments{
\item{xx.0}{True hidden sequence}

\item{xx.Vit}{Hidden sequence estimated by Viterbi}

\item{par}{Model parameters}

\item{opts}{A list of options for the estimation. Default is \code{opts =
list(n.seeds = 3, d0 = 3, SS = NULL, rotate = FALSE)} where \code{n.seeds} is
the number of seeds for the optimistic search, \code{d0} is the size of the
smallest search interval, \code{SS} is the initial partition and
\code{rotate} indicates whether or not the gain functions have to be rotated.}
}
\value{
Estimated sequence and its probability
}
\description{
Quick Adaptive Ternary Segmentation, with display
}
