% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QATS.R
\name{QATS.R}
\alias{QATS.R}
\title{Quick Adaptive Ternary Segmentation, R version}
\usage{
QATS.R(par, opts = list(n.seeds = 3, d0 = 3, SS = NULL, rotate = FALSE))
}
\arguments{
\item{par}{Model parameters}

\item{opts}{A list of options for the estimation. Default is \code{opts =
list(n.seeds = 3, d0 = 3, SS = NULL, rotate = FALSE)} where \code{n.seeds} is
the number of seeds for the optimistic search, \code{d0} is the size of the
smallest search interval, \code{SS} is the initial partition and
\code{rotate} indicates whether or not the gain functions have to be rotated.}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
Quick Adaptive Ternary Segmentation, R version
}
