% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiors.R
\name{Viterbi.CPP}
\alias{Viterbi.CPP}
\title{Viterbi decoder, C++ version}
\usage{
Viterbi.CPP(par, n.rep = 1)
}
\arguments{
\item{par}{Model parameters}

\item{n.rep}{Number of repetition (for timing)}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
Viterbi decoder, C++ version
}
