% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiors.R
\name{Viterbi.R}
\alias{Viterbi.R}
\title{Viterbi decoder, R version}
\usage{
Viterbi.R(par)
}
\arguments{
\item{par}{Model parameters}
}
\value{
Estimated sequence, its probability and estimation time
}
\description{
Viterbi decoder, R version
}
